there is a bug from flow for collecting

+++ app/src/main/java/com/androidint/todo/repository/model/TaskWithCategoryName.kt
@@ -7,7 +7,7 @@ data class TaskWithCategoryName(
     val description: String?,
     @Embedded
     val day: Day,
-    val ownerCategoryId: Int,
+    val ownerCategoryId: Long,
     @Embedded
     val timeDuration: TimeTask,
     val priority: Int = 1,
